refactor: 사용되지 않는 메세지 제거
